fix(dispatcher): fix error when multiple tools calls
Fix the bug where the prefix of the call information was not correctly
cleared when there were multiple tools calls.

Fix an error due to a typo in judging the \" character.

Fix the error of only searching once for the to_user_content_end end marker
in the buffer. It should scan all double quotes, not just the first one.